fixed typo in sim_fake_users

+++ emission/simulation/simulate_fake_users.py
@@ -53,7 +53,7 @@ def main (usercount, tripcount):
         "radius" : ".1"
     }
 
-    for i in range((usercount / 50) + 1):
+    for i in range(int(usercount / 50) + 1):
         curr_num_users = (usercount - 50 * i) % 51
         fakeusers = create_fake_users (curr_num_users, base_user_config, client_config) 
         create_and_sync_data (fakeusers, tripcount)
